Always put the output in `viewer` for ease of use
